fixes tests
