+++ client/tests/App.test.tsx
@@ -1,8 +1,14 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import * as React from 'react';
+import {
+  it, expect, describe, vi,
+} from 'vitest';
 import App from '../src/components/App';
 import MockProvider from './MockProvider';
+
+vi.mock('../src/utils', () => ({
+  getTasks: vi.fn(),
+}));
 
 describe('App', () => {
   it('renders App component', () => {
